nvim: remove chromatica due to CursorMoved and LSP
Two issues that cause me to remove Chromatica, as well as one underlying
reason for that experience.

Firstly, Chromatica has an augroup called "chromatica" on every
CursorMoved that, although I'm not sure, seemed to have an effect on
other buffers. Generally I'd prefer that such groups only activate in
buffers belonging to the same file type as it supports.

Secondly, I'm using ALE and considering trying nvim-coc to get full LSP
support underneath. In either case, Chromatica's value is limited, but
particularly so when compared to nvim-coc.