zsh: add brew helpers for sourcing from multiples
Compared to earlier I need to handle two very different sets of file
structures. I suppose we could set something up in a different way,
perhaps if we could all run it in a different file structure or a chroot
similar to it, but it's not.

Thus, add two simple functions for either sourcing the first from a list
and a little brew helper.

(I'm not a fan of Homebrew packages. It feels messy.)