feat(nvim): use lazy plugin manager & move lua config
Move the two lua files that formerly resided in the sections folder.
The move to Lazy stemmed from the maintainer of Packer themself having
moved to Lazy, which sort of signals the chances of Packer remaining
up-to-date. Additionally, it features better loading, and doesn't rely
on the built-in package management in neovim (although there's nothing
inherently wrong with it).

However, it does do something different from the compilation step that
Packer does, and which I experienced multiple issues with when any
configuration for a plugin was incorrect, leading to "locking myself
out" of nvim unless I --noplugin and fix the error. Annoying.
This resolves that, allowing for more control with an even better user
interface.