refactor(nvim): move fzf and vimtex into lazy
Separates fzf from vimtex and disables vimtex as I'm not actively using
it as much as I could be. vimtex can be enabled at a later point. I'll
need to fix the @@ citation keybinding at the same time.

fzf has a slightly improved setup with lazy handling the key management
(nice), and I've removed the redefinitions for :Files and :Rg because
they were equivalent or identical to the default definitions.