misc(poetry): unset VIRTUAL_ENV & format
Noticed a problem with mise and direnv/poetry whereby certain projects
would seemingly get stuck in their own state. Almost as if the
VIRTUAL_ENV was still set while we were evaluating whether there was a
virtual environment there or not.

Proceeded to unset the value before proceeding, which seemed to work. A
sub-shell might've been a cleaner approach to the problem.